requirements.txt
openai
requests
tabulate
python-dotenv